починенное время
отдельно рабочий день, отдельно задачи
пересчет переработок, недоработок, бонусов и зп

+++ RemoteControl1/Models/ActivityLog.cs
@@ -25,7 +25,7 @@ namespace RemoteControl1.Models
 
         public DateTime? EndedAtUtc { get; set; }
 
-        public double DurationHours { get; set; }
+        public decimal DurationHours { get; set; }
             
         public string? Comment { get; set; }
